+++ src/lib/T3000/Hvac/App/App.ts
@@ -1,5 +1,5 @@
-import HvacModels from "../Hvac.Models";
-import * as Utils from '../Hvac.Utils';
+import HvacModels from "../Model/Hvac.Models";
+import * as Utils from '../Helper/Hvac.Utils';
 import { cloneDeep } from "lodash";
 import Doc from "../Document/DocHandler";
 import UI from "../UI/UI";
