+++ src/lib/T3000/Hvac/Helper/Hvac.SVG.js
