+++ src/lib/T3000/Hvac/Helper/Hvac.SVG1.ts
