+++ src/lib/T3000/Hvac/Hvac.ts
@@ -1,7 +1,7 @@
 
 import App from './App/App';
-import * as Utils from './Hvac.Utils';
-import Models from './Hvac.Models';
+import * as Utils from './Helper/Hvac.Utils';
+import Models from './Model/Hvac.Models';
 import Doc from './Document/DocHandler';
 import UI from './UI/UI';
